Drop the Condition abstraction (#1956)
Drop the `Condition` interface and replace with callbacks. The `it()`
utility can cause confusion and look like a language feature. Another
potential benefit of function types is they may get syntax or formatting
improvements in the future.

One downside of the change is when inference fails, for instance when
passing a callback in a collection for `deepEquals`, there is no static
help. At runtime there will be errors about `Subject` not having
instance methods matching the extensions that are used.

One upside is some linting setups may be able to statically help avoid
passing a `Future<void> Function(Subject)` where a synchronous condition
is required. This static checking couldn't work for the `it()` syntax.

Tests are mostly migrated to use `it` as the subject name for
convenience of the migration, but tests in the real world (and in the
example code in docs) may be more likely to choose a more meaningful
letter, or a longer name based on the value.

Add special handling in `literal` for `void Function(Subject)` instances
since there is no longer a meaningful `toString()` implementation on the
instances.