Timeout browser suite loads (#2080)
The current timeout applies to starting the process and connecting back
from the host. Add a timeout which applies on each suite as it is
loaded.

Change the chrome argument to log to `stderr` instead of `stdout`. The
latter seems to have no effect.

Eagerly decode and split lines from the browser output instead of
storing the bytes to avoid mixing lines between the two outputs. In most
cases this shouldn't matter, I have not seen output on `stdout`.

Use `--v=0` instead if `--v=1` since the 0 level includes console
output, and the extra browser level output with higher verbosity does
not look useful.