Add some console logging to browser test startup (#2083)
Add support for console log and warn level in the dom library.
Switch an existing `print` to a `console.warn` for an error. Use
`console.log` consistently otherwise. This is less important in the host
(where a `print` will anyways log to the console) than it is in the
remote listener (where a `print` could be intercepted and forwarded to
the test runner).

Refactor an `if`/`else` chain over a map field to a `switch` with
destructuring for all the patterns that can be handled. Add a default
case with a console warning instead of silently treating potentially
non-matching commands as `'closeSuite'`. This also changes some
potential cast failures into falling through to the warning case which
should be easier to debug.

Add a message from the frame when it is connecting back to the host.