Extend the timeout for runtime_matrix_test (#2084)
This test invokes the test runner in each case, and when the outer test
timeout is faster than a timeout within the sub-process it prevents us
from getting potentially useful output. Double the outer timeout so that
any timeouts from the sub-process runner have a better chance at firing
and emitting output before the test is killed.