Make builder not merge generic extension.
The previous behaviour was that if the user adds a build_extension,
the build would inject the generic build extension to cover all files.
This was useful to simplify the build.yaml file at the cost of
unexpected behaviour.

To avoid potencial unexpected behaviour,
if the user provides custom build_extensions we assume that the patterns
cover all files and therefore we do not merge the generic
build_extension.