Add a silent reporter (#2093)
Resolve a flaky test by eliminating the reporter output as a
potential race condition.

Add a `hidden` flag to `ReporterDetails` since the reporter is not worth
showing in the CLI usage help. Request the silent reporter for a test
which is flaky due to interleaving of forwarded `stdout` with the
reporter's output. During flakes the `hello` and the newline can come as
separate events in the stream, so sometimes the newline is written
following unexpected reporter content.