Remove unused href field from error report
Originally the href was added manually to the data to work around a
`dart:html` limitation. A different workaround for the host side was
implemented in #2066, but the frame side was not changed in the same
commit to allow some cross-compatibility. Remove the final manual href
field now that no host implementations would read it.