Strip leading slash when spawning isolate
Pull out a small utility for stripping leading slashes on windows from
the arg parsing code so it can be reused when spawning isolates by path.

It's still not clear what specific conditions lead to paths which look
like this, but the behavior is triggering for this path in another CI
system.