Use package: URIs for imports when possible
Fixes #2142

When a compiler is reading imports it's possible for a relative file
import and absolute package import to the same library to get treated as
separate libraries. The VM works around this in the case of a `main`
under `lib/` by checking when the entrypoint, and only the entrypoint,
could be referenced with a `package:` URI so that further relative URIs
resolved from that point are consistently treated as `package:` URIs.

Use the same workaround for bootstrap files. Add an `absoluteUri`
utility in `test_core` which prefers `package:` URIs when feasible, and
falls back to absolute file URIs otherwise. When this initial import
uses a `package:` URI it will make the subsequent relative imports
correctly resolve as `package:` URIs.

Add `async` and `await` as necessary since we're currently using the
async APIs to read the package config.