Remove support for Internet Explorer (#2114)
Closes #1614

Use should be replaced by the edge browser.

It has been 2 years since we removed support for internet explorer in
dart2js. We have not been testing internet explorer integration, and it
may already be broken by dart2js changes.

Remove every reference to `ie` or internet explorer in the repo.

Treat this as non-breaking for `package:test` even though there is a
small risk of breaking some users still using internet explorer and it
happens to still be working.