Document the silent reporter (#2163)
Towards #2162

Unhide the silent reporter. This was originally introduced hidden
because we expected it to only be useful for our own CI infrastructure,
since a failure without any output cannot be investigated - it must be
reproduced in a separate run to even find which test case failed.

We now have an external request for failing without any output on stdio
and there is no strong reason to keep the silent reporter hidden.

Remove the support for hidden reporters entirely - it can easily be
reintroduced if we have another reporter which isn't user facing.

Expand the reporter description. Break a long description string into a
adjacent literals to avoid lines over 80 columns.