Try letting the IsolateChannel close the RecievePort
The `IsolateChannel` should close the receive port when we call
`outerChannel.sink.close` which is also added to the `cleanupCallbacks`.

This was originally added in #1280 to try to fix cases where the test
runner could hang after running some tests specifically on the test
runner CI when running the test as a subprocess. That PR went through
changes closing a number of ports speculatively, and this was one likely
not necessary.