Let the IsolateChannel close the ReceivePort (#2196)
The `IsolateChannel` should close the receive port when we call
`outerChannel.sink.close` which is also added to the `cleanupCallbacks`.

This was originally added in #1280 to try to fix cases where the test
runner could hang after running some tests specifically on the test
runner CI when running the test as a subprocess. It looks like when
it was introduced there was no call to `outerChannel.sink.close()`.
A cleanup callback to close the sink was added in #1941 which made
this unnecessary.