Add tests
Add a unit test for the described behavior.
Update an integration test to demonstrate that a suite level `@Skip()`
annotation cannot be overridden, it will always show as _one_ test
skipped.

Remove all but the first `[]` link for each reference in the paragraph.
Surround the other references with backticks for code blocks.