Try removing isolate.kill()
Originally isolates were killed when the loader was closed as the runner
is shutting down. There was no comment in the review about the specific
reason for `isolate.kill()`. https://codereview.chromium.org//920703006

Later the kill was moved to when the test suite is done. There was no
specific comment about the `isolate.kill()` changes.
https://codereview.chromium.org//1206033004

No internal tests seem to be relying on this behavior, and nothing times
out when it is removed.

Check whether there are any tests on CI that validate this behavior.