Review changes
Use `packageConfigUri` - we should already be validating a non-null
package config for platforms where this is used. Use a non-null
argument.

Don't reuse part of the args in the template - repeat the arg name in
`mainArgs` and `forwardedArgName`.

Add a comment on the variable to remind us that usage will not be
obvious.