Add package config URI into VM test entrypoints (#2245)
Towards #2246 

This will allow a client of the VM service to evaluate the top level
variable and locate the package config used by the test runner.

Add a public `testBootstapContents` function in `test_compiler.dart`
which centralizes the bootstrap templates. Add an argument for the
package config URI and a `VmTestType` which is either "Vm" or
"Native".

Add the first test case in `test_core`. Add the pub dependency and
override for `package:test`, as well as the mono_repo configuration to
run unit tests. Other tests for `test_core` code is under the `test`
package but should be moved eventually.