Enable avoid_dynamic_calls lint (#2410)
Add argument types to `typedMatches` arguments which were unnecessarily
widening it back to `dynamic`. 

Add some missing `as dynamic` and `as Function` which is the pattern the
lint uses to allow intentional dynamic calls.

Ignore some dynamic calls in a test for a legacy API.

Ignore a couple false positives for calling methods (as opposed to
getters) on a cast expression. The false positive was fixed in
https://dart-review.googlesource.com/c/sdk/+/390640
but not yet published.

Remove the deprecated lint `package_api_docs`.

Replaces https://github.com/dart-lang/matcher/pull/205