Prepare mockito for the 'wildcard-variables' language feature.
We generally preserve positional parameter names, but in the case of wildcard
parameters, we must rename them. We cannot override a wildcard parameter with a
wildcard parameter, since we do use the parameter (we pass it to
`super.noSuchMethod` in a call to `Invocation.method`). This change introduces
some renaming logic for wildcard parameters, using new method-unique parameter
names.

PiperOrigin-RevId: 696158750