Resolve for typedefs/aliases and records in typeArguments & resolve duplicate typedefs
https://github.com/dart-lang/mockito/pull/776

Closes: https://github.com/dart-lang/mockito/issues/775
Closes: https://github.com/dart-lang/mockito/issues/777
Closes: https://github.com/dart-lang/mockito/issues/778

The fix for these issues is closely related so I've done both of them in a single PR.

- Mockito wasn't checking for typedefs/aliases in `typeArguments`
- it also wasn't looking into records on type args.
- Typedefs which had the same underlying type were ignored even if the typedef had a different name

These issues cause the dreaded `is missing from the asset URI mapping` error.

See the related issue for a repro of each of these.

I've also added regression tests for all of these.

PiperOrigin-RevId: 702712435