Use dart_flutter_team_lints for matcher (#2426)
Change the dependency and analysis options include to the stricter team
configuration.

Fix newly surfaced diagnostics. Most code changes are related to the
strict modes that are enabled by the team config.
Fix most implicit casts with explicit casts, but fix some arguments in
`typedMatches` implementation so they don't unnecessarily widen the type
back to `dynamic` before using the value.

Bump to Dart 3.5 The min SDK for the test runner is already at 3.5.0 and
matcher is only usable with the latest version of the test runner.