Make zone handlers call zone.run to run callbacks.
The platform library `Zone` implementation will change to
not use `bindCallbackGuarded` in the `Timer` constructors
and top-level `scheduleMicrotask` function.
Instead it only calls `registerCallback`, and then the
`Zone.create{,Periodic}Timer` and `Zone.scheduleMicrotask`
are in charge of using `Zone.runGuarded` on the callback
when timer/microtask event happens.

This ensures that a surrounding zone's `registerCallback`
can't make the callback throw in a way that is outside
of the `runGuarded` call.