Allow specifying an explicit location for test/groups
Usually test locations (which are used both to report the location of tests in the json reporter, and also for filtering tests by line/col) are inferred from the call stack when test() is called.

This changes adds a new `location` parameter to group/test to allow this information to be provided explicitly. This can be useful where the call to `test()` doesn't contain the actual location of the declaration in the call stack (for example when using `pkg:test_reflective_loader`).

See https://github.com/dart-lang/test/issues/2029
See https://github.com/Dart-Code/Dart-Code/issues/5480