Fix an issue with failed assertions using setUpAll/tearDownAll (#2499)
When groups are filtered using `filter()` and not `onPlatform`, we would clone the group and tests without also cloning `setUpAll`/`tearDownAll`, which led to an assertion failure when trying to set the parents.

This change ensures we clone tests in `filter()` like we do in `forPlatform()`.

Fixes https://github.com/dart-lang/test/issues/2498