Consider NaN a match for itself
Closes #2520

A `double.nan` is not equal to itself with the `==` operator, but for
the purposes of testing it's easier to consider a `NaN` correct where a
`NaN` was expected. This matches behavior in several other testing
frameworks in other languages. This is unlikely to cause existing tests
to fail or change semantics since a test is unlikely to use non-equality
as a positive indication of a `NaN` result.