Allow connecting to the runner from test suite
Closes #2497

Add a zone variable when the declarer runs `main` from the test suite
which gives access to the `connectOut` callback which creates a channel
back to the runner with a provided name.

Add an optional argument allowing a map of zone values to the `declare`
method and use it from the remote listener.