Drop executable arguments forwarding
Closes #2527

During the null safety migration there were `dart` level flags that
needed to be kept in sync for precompiled JS tests so the platform
executable arguments were forwarded direclty. This is currently causing
a test failure, and there are no null safety related `dart` level flags
so drop the forwarding.