Remove extra version info
Follow up to #2538

This is only relevant when using the test runner as a git or path
dependency. In these cases the version will already end in `-WIP` with
out current development practices and in the places where it may come up
the users don't need help understanding where the code is coming from.