Fix stdout color support check
Currently `canUseSpecialChars` returns `true` when not in Windows and
not in tests of the package itself.

This causes color codes to be printed when stdout is redirected or
piped.

Fix it by checking if stdout is a tty. If it is, then we assume that
most terminals today support colors and return `true`.

(We can't use the standard library `supportsAnsiEscapes` as it only
returns `true` when `$TERM` is `xterm`, see
https://github.com/dart-lang/sdk/issues/31606.)

Also fix the extended reporter's color support flag initialization,
which was hard-coded as `true`.

Fixes #19.