Make all package:coverage imports conditional (#2544)
The `package:coverage` library imports `dart:io` directly and cannot be
a dependency of apps compiled to platforms without `dart:io`. All
imports and uses of `package:coverage` must be through platform specific
imports. Add extra indirection in the `coverage.dart` and
`coverage_stub.dart` libraries so that `engine.dart` does not need to
import `package:coverage` or refer to the `HitMap` type it exports. Add
a forwarding implementation of the `merge` extension.