Require a definition named `main` in test suites
Closes #2567

The error when there is a compilation failure due to a missing `main` is
not very clear. Add an explicit check for a definition with that name
when parsing metadata.

This does not handle cases where the definition is not a method, but
those cases should be more rare than omitting main.