Require a definition named `main` in test suites (#2569)
Closes #2567

The error when there is a compilation failure due to a missing `main` is
less direct than it could be and exposes details of wrapping the test
suite in another library. Add an explicit check for a function declared
with that name when parsing metadata before attempting any compiles.

Add a check to `parseMetadata` that throws a `FormatException` if there
is no function declared directly in the library named `main`. This will
prevent some previously working designs such as exposing a main in a
part or with `export`, but we do not expect those designs are used in
practice.